requirements.txt
fastapi==0.111.1
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
email-validator==2.2.0